Added DelayedConfirmationView to give a sense of how long we battle. Automatically moves to the next activity which is battle results.

+++ LongJourney/LJ Watch/src/main/java/com/lchrislee/longjourney/activities/BattleSneakActivity.java
@@ -56,7 +56,7 @@ public class BattleSneakActivity extends Activity {
         @Override
         public void onClick(View v) {
             Intent nextActivity;
-            if(BattleUtility.determineSneakSuccess(v.getTag())){
+            if(BattleUtility.determineSneakSuccess((Integer) v.getTag())){
                 nextActivity = new Intent(v.getContext(), SpoilsActivity.class);
                 nextActivity.putExtra(SpoilsActivity.CONCLUSION, BattleUtility.BATTLE_OPTION_SNEAK);
             }else{
